Extract a Solr client.
Pulling all the Solr interactions out to a client should make it easier
to work on things like #521 and standardize our return values.
